Agregado de repositories imagenes, models imagenes, métodos en imagen services, cambios en schema y agregado de imagenes en el FrontEnd

+++ TPO/Api/src/main/java/com/uade/api/models/ServicioModel.java
@@ -16,7 +16,7 @@ import java.util.List;
 public class ServicioModel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idServicio;
+    private Long idServicio;
     @ManyToOne
     @JoinColumn(name="documentoVecino")
     private VecinoModel vecino;
